Converting more handlers to new format
