+++ old/app/device_handler.py
