+++ old/ls336/temp_336_ioc.py
